refactor: move repository files to db folder
