+++ src/data/usecases/add-account/db-add-account-protocols.ts
@@ -2,4 +2,4 @@
 export * from '../../../domain/usecases/add-account'
 export * from '../../../domain/models/account'
 export * from '../../protocols/criptography/encrypter'
-export * from '../../protocols/add-account-repository'
+export * from '../../protocols/db/add-account-repository'
