+++ src/main/decorators/log.spec.ts
@@ -1,7 +1,7 @@
 import { HttpRequest, HttpResponse, Controller } from '../../presentation/protocols'
 import { serverError } from '../../presentation/helpers/http/http-helper'
 import { LogControllerDecorator } from './log'
-import { LogErrorRepository } from '../../data/protocols/log-error-repository'
+import { LogErrorRepository } from '../../data/protocols/db/log-error-repository'
 
 const makeLogErrorRepository = (): LogErrorRepository => {
   class LogErrorRepositoryStub implements LogErrorRepository {
